requirements.txt
flask>=3.1.1
flask-sqlalchemy>=3.1.1
gunicorn>=23.0.0
matplotlib>=3.10.3
numpy>=2.3.0
openai>=1.88.0
opencv-python>=4.11.0.86
pandas>=2.3.0
pillow>=11.2.1
psycopg2-binary>=2.9.10
scikit-learn>=1.7.0
seaborn>=0.13.2
torch>=2.7.1
torchvision>=0.22.1
werkzeug>=3.1.3
email-validator>=2.2.0 